--- app/src/main/java/com/example/movieswatchnow/core/MovieDatabase.kt
@@ -1,60 +1,0 @@
-package com.example.movieswatchnow.core
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import androidx.sqlite.db.SupportSQLiteDatabase
-import com.example.movieswatchnow.domain.Movie
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-
-@Database(
-    entities = [Movie::class],
-    version = 1
-)
-abstract class MovieDatabase : RoomDatabase() {
-    abstract fun dao(): MovieDao
-
-    private class MovieDatabaseCallback(
-        private val scope: CoroutineScope
-    ) : RoomDatabase.Callback() {
-
-        override fun onCreate(db: SupportSQLiteDatabase) {
-            super.onCreate(db)
-            INSTANCE?.let { database ->
-                scope.launch {
-                    val wordDao = database.dao()
-
-
-                    // Add sample words.
-                    val word = Movie(id = 1, title = "Hello")
-                    wordDao.saveMovies(word)
-
-                }
-            }
-        }
-    }
-
-    companion object {
-        // Singleton prevents multiple instances of database opening at the
-        // same time.
-        @Volatile
-        private var INSTANCE: MovieDatabase? = null
-
-        fun getDatabase(context: Context): MovieDatabase {
-            // if the INSTANCE is not null, then return it,
-            // if it is, then create the database
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    MovieDatabase::class.java,
-                    "movie_database"
-                ).build()
-                INSTANCE = instance
-                // return instance
-                instance
-            }
-        }
-    }
-}
