--- app/src/main/java/com/example/movieswatchnow/core/extensions/Extensions.kt
@@ -1,18 +1,0 @@
-package com.example.movieswatchnow.core.extensions
-
-import android.view.LayoutInflater
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import androidx.viewbinding.ViewBinding
-import com.example.movieswatchnow.core.ui.FragmentViewBindingDelegate
-
-
-inline fun <T : ViewBinding> AppCompatActivity.viewBinding(
-    crossinline bindingInflater: (LayoutInflater) -> T,
-) = lazy(LazyThreadSafetyMode.NONE) {
-    bindingInflater.invoke(layoutInflater)
-}
-
-fun <T : ViewBinding> Fragment.viewBinding(viewBindingFactory: (View) -> T) =
-    FragmentViewBindingDelegate(this, viewBindingFactory)
