--- app/src/main/java/com/example/movieswatchnow/presentation/adapter/MovieDetailsAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.movieswatchnow.presentation.adapter
-
-class MovieDetailsAdapter {
-}
