--- app/src/main/java/com/example/movieswatchnow/presentation/viewmodel/BaseViewModel.kt
@@ -1,14 +1,0 @@
-package com.example.movieswatchnow.presentation.viewmodel
-
-import androidx.lifecycle.ViewModel
-import io.reactivex.disposables.CompositeDisposable
-
-open class BaseViewModel : ViewModel() {
-
-    protected val compositeDisposable = CompositeDisposable()
-
-    override fun onCleared() {
-        super.onCleared()
-        compositeDisposable.clear()
-    }
-}
