+++ app/src/main/java/com/myapp/movieswatchnow/core/ui/FragmentViewBindingDelegate.kt
@@ -1,0 +1,52 @@
+package com.myapp.movieswatchnow.core.ui
+
+import android.view.View
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.DefaultLifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.Observer
+import androidx.viewbinding.ViewBinding
+import kotlin.properties.ReadOnlyProperty
+import kotlin.reflect.KProperty
+
+class FragmentViewBindingDelegate<T : ViewBinding>(
+    val fragment: Fragment,
+    val viewBindingFactory: (View) -> T,
+) : ReadOnlyProperty<Fragment, T> {
+
+    private var binding: T? = null
+
+    init {
+        fragment.lifecycle.addObserver(object : DefaultLifecycleObserver {
+            val viewLifecycleOwnerLiveDataObserver =
+                Observer<LifecycleOwner?> {
+                    val viewLifecycleOwner = it ?: return@Observer
+                    viewLifecycleOwner.lifecycle.addObserver(object : DefaultLifecycleObserver {
+                        override fun onDestroy(owner: LifecycleOwner) {
+                            binding = null
+                        }
+                    })
+                }
+
+            override fun onCreate(owner: LifecycleOwner) =
+                fragment.viewLifecycleOwnerLiveData.observeForever(
+                    viewLifecycleOwnerLiveDataObserver)
+
+            override fun onDestroy(owner: LifecycleOwner) =
+                fragment.viewLifecycleOwnerLiveData.removeObserver(
+                    viewLifecycleOwnerLiveDataObserver)
+        })
+    }
+
+    override fun getValue(thisRef: Fragment, property: KProperty<*>): T {
+        val binding = binding
+        if (binding != null && binding.root === thisRef.view) { return binding }
+        val view = thisRef.view ?: throw IllegalStateException(fragmentViewBindingError)
+        return viewBindingFactory(view).also { this.binding = it }
+    }
+
+    private companion object {
+        private const val fragmentViewBindingError =
+            "Should not attempt to get bindings when Fragment views are destroyed."
+    }
+}
