+++ app/src/main/java/com/myapp/movieswatchnow/presentation/adapter/MovieDetailsAdapter.kt
@@ -1,0 +1,4 @@
+package com.myapp.movieswatchnow.presentation.adapter
+
+class MovieDetailsAdapter {
+}
