+++ app/src/main/java/com/myapp/movieswatchnow/presentation/viewmodel/BaseViewModel.kt
@@ -1,0 +1,14 @@
+package com.myapp.movieswatchnow.presentation.viewmodel
+
+import androidx.lifecycle.ViewModel
+import io.reactivex.disposables.CompositeDisposable
+
+open class BaseViewModel : ViewModel() {
+
+    protected val compositeDisposable = CompositeDisposable()
+
+    override fun onCleared() {
+        super.onCleared()
+        compositeDisposable.clear()
+    }
+}
