+++ app/src/main/res/navigation/movies_nav_graph.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/moviesListFragment"
-        android:name="com.example.movieswatchnow.presentation.fragment.MovieListFragment"
+        android:name="com.myapp.movieswatchnow.presentation.fragment.MovieListFragment"
         android:label="Movies List"
         tools:layout="@layout/fragment_movies_list" />
 
     <fragment
         android:id="@+id/moviesSavedFragment"
-        android:name="com.example.movieswatchnow.presentation.fragment.MovieSavedFragment"
+        android:name="com.myapp.movieswatchnow.presentation.fragment.MovieSavedFragment"
         android:label="Movies Saved List"
         tools:layout="@layout/fragment_movies_list" />
 
     <fragment
         android:id="@+id/moviesDetailsFragment"
-        android:name="com.example.movieswatchnow.presentation.fragment.MovieDetailsFragment"
+        android:name="com.myapp.movieswatchnow.presentation.fragment.MovieDetailsFragment"
         android:label="Movies Details"
         tools:layout="@layout/fragment_movie_detail">
         <action
